New Landing page for non users
- Edited a video for this new background
