+++ client/src/pages/Home.tsx
@@ -1,74 +1,60 @@
-import { Text, Input, Tr, Td, Button, ChakraProvider, Tbody, Grid, GridItem, HStack, TableContainer, Accordion, AccordionItem } from "@chakra-ui/react";
+import React, { useEffect, useState } from "react";
+import { ChakraProvider, Box, Text } from "@chakra-ui/react";
 import Header from "../components/Header";
-import { useEffect, useState } from "react";
-import { start } from "repl";
-import { CloseIcon, Icon } from "@chakra-ui/icons";
 
 function Home() {
-  const [task, setTask] = useState<string>('');
-  const [tasks, setTasks] = useState<string[]>([]);
-
+  const [isShow, setIsShow] = useState<boolean>(true)
   useEffect(() => {
-    const savedTasks = localStorage.getItem('tasks');
-    if (savedTasks) {
-      setTasks(JSON.parse(savedTasks));
-    }
+    const check = localStorage.getItem("user");
+    setIsShow(check === "false");
   }, []);
-
-  useEffect(() => {
-    localStorage.setItem('tasks', JSON.stringify(tasks));
-  }, [tasks]);
-
-  const addTask = () => {
-    if (task.trim() !== '') {
-      setTasks([...tasks, task]);
-      setTask('');
-    }
-  };
-
-  const deleteTask = (index: number) => {
-    const updatedTasks = tasks.filter((_, i) => i !== index);
-    setTasks(updatedTasks);
-  };
-
   return (
     <ChakraProvider>
       <Header />
-      <Text></Text>
-      <Grid templateColumns="1fr 1fr" gap={4}>
-        <GridItem>
-          <HStack>
-            <Input
-              type="text"
-              color={"black"}
-              placeholder="Add a new task"
-              value={task}
-              onChange={(e) => setTask(e.target.value)}
-            />
-            <Button colorScheme="blue" onClick={addTask}>
-              Add
-            </Button>
-          </HStack>
-        </GridItem>
-        <GridItem>
-          <Accordion>
-            {tasks.map((task, index) => (
-              <AccordionItem key={index} w="100%" justifyContent="space-between" alignItems="center">
-                <Text color={"black"} p={1} m={3}>{task}
-                <Button
-                  colorScheme="red"
-                  size="sm"
-                  marginLeft={3}
-                  onClick={() => deleteTask(index)}
-                >
-                  <CloseIcon/>
-                </Button>
-                </Text>
-              </AccordionItem>
-            ))}
-          </Accordion>
-        </GridItem>
-      </Grid>
+      <Box position="relative">
+      {isShow && <div style={{ position: "relative", width: "100%", height: "0px", paddingBottom: "56.250%" }}>
+        <iframe
+          allow="fullscreen;autoplay"
+          allowFullScreen
+          height="100%"
+          src="https://staging.streamable.com/e/kk68s5?autoplay=1&muted=1&nocontrols=1"
+          width="100%"
+          style={{
+            border: "none",
+            width: "100%",
+            height: "100%",
+            position: "absolute",
+            left: "0px",
+            top: "0px",
+            overflow: "hidden",
+            // backgroundColor: "rgba(0, 0, 0, 0.5)",
+          }}
+          ></iframe>
+          <div
+            style={{
+              position: "absolute",
+              top: 0,
+              left: 0,
+              width: "100%",
+              height: "100%",
+              backgroundColor: "rgba(0, 0, 0, 0.5)", // Semi-transparent black overlay
+              zIndex: 1,
+            }}
+            ></div>
+        </div>}
+        <Box
+          position="absolute"
+          top="20%"
+          left="30%"
+          transform="translate(-50%, -50%)"
+          textAlign="center"
+          zIndex={2}
+        >
+          <Text fontSize="5xl" color="white" fontWeight="bold" fontFamily={"arial"}>
+            welcome to algovisor
+          </Text>
+        </Box>
+      </Box>
     </ChakraProvider>
   );
 }
